requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
python-dotenv
